Solution to Check if a Number is Square of Another

+++ Check if a Number is Square of Another@@6788dd63490637cfc44e1ba4/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main()
 {
-    float a,b;
-    scanf("%f %f", &a, &b);
+    double a,b;
+    scanf("%lf %lf", &a, &b);
     if (a=b*b){
         printf("Yes");
     }
